Update and rename georgiahl31jan.cpp to competition.cpp
This program was created for a fictional company that runs competitions. The task was to create a function to enable the company to enter data (name and score) for the three finalists in a competition (first place, second place and third place). The n range is between 0 and 1000. If a finalist scores between 900 and 1000, they are the winner; if their score is between 800 and 900, they are second place; and if their score is between 700 and 800, they are third place. The program uses a while loop which runs until three finalists and three valid scores have been entered; if the user enters a score that is under 700 or above 1000, the program throws up an error message and they are asked to re-enter a valid score. Once the conditions of the while loop are met, the program outputs the name of each of the finalists and their score, in order of first place, second place, third place.